[rename] components - main 폴더 생성, Carousel.js, Preview.js 이동
